Padronizado em C. Add Makefile.
